Revert changes to file structure
